closes #424
